+++ src/services/ReportsService.php
@@ -33,6 +33,21 @@ class ReportsService extends Component
 	 */
 	public $debug = false;
 
+	/**
+	 * The array of filter types and their respective search syntax.
+	 * @var array
+	 */
+	const FILTER_TYPES = [
+		'contains' => '*[value]*',
+		'starts with' => '[value]*',
+		'ends with' => '*[value]',
+		'is equal to' => '[value]',
+		'is greater than' => '> [value]',
+		'is less than' => '< [value]',
+		'is empty' => ':empty:',
+		'is not empty' => ':notempty:'
+	];
+
 	public function init()
 	{
 		$this->plugin = LinkVault::getInstance();
@@ -90,103 +105,122 @@ class ReportsService extends Component
 	}
 
 	/**
-	 * This method cleans up a download record array by reference so odd criteria
-	 * attributes aren't displayed or taking up space in memory.
-	 * @param array $r
+	 * This method generates a single piece of element criteria for a given
+	 * field handle, filter type and optional value.
+	 * @param string $fieldHandle,
+	 * @param string $filterType
+	 * @param mixed $value
+	 * @return array
 	 */
-	public function cleanRecordArray(&$r)
+	public function fieldCriteria($fieldHandle, $filterType, $value=null): array
 	{
-		unset($r['title'],
-			$r['slug'],
-			$r['uri'],
-			$r['before'],
-			$r['after'],
-			$r['tempId'],
-			$r['fieldLayoutId'],
-			$r['contentId'],
-			$r['hasDescendants'],
-			$r['ref'],
-			$r['structureId'],
-			$r['totalDescendants'],
-			$r['newSiteIds'],
-			$r['archived'],
-			$r['enabled'],
-			$r['draftId'],
-			$r['revisionId'],
-			$r['dateDeleted'],
-			$r['trashed'],
-			$r['propogateAll'],
-			$r['resaving'],
-			$r['duplicateOf'],
-			$r['previewing'],
-			$r['hardDelete']
-		);
+		if ( $fieldHandle && $filterType ) {
+			$criteria = [$fieldHandle => str_replace('[value]', $value, self::FILTER_TYPES[$filterType])];
+		} else {
+			$criteria = [];
+		}
+		return $criteria;
 	}
 
 	/**
-	 * This method returns an associative array of LinkVaultDownloadQuery criteria
-	 * attributes and their respective option label.
+	 * This method converts filter input criteria into query criteria.
+	 * @param array $input
 	 * @return array
 	 */
-	public function reportAttributeOptions(): array
+	public function formatCriteria($input): array
 	{
-		/*
-		This is a long list of criteria that don't apply to Link Vault downloads.
-		This list seems to get longer with each release of Craft.
-		*/
-		$omittedCriteria = [
-			'after',
-			'ancestorDist',
-			'ancestorOf',
-			'archived',
-			'before',
-			'descendantDist',
-			'descendantOf',
-			'draftCreator',
-			'draftId',
-			'draftOf',
-			'drafts',
-			'enabledForSite',
-			'fixedOrder',
-			'hasDescendants',
-			'ignorePlaceholders',
-			'inReverse',
-			'leaves',
-			'level',
-			'nextSiblingOf',
-			'newSiteIds',
-			'preferSites',
-			'prevSiblingOf',
-			'positionedAfter',
-			'positionedBefore',
-			'orderBy',
-			'ref',
-			'relatedTo',
-			'revisionCreator',
-			'revisionId',
-			'revisionOf',
-			'revisions',
-			'search',
-			'siblingOf',
-			'status',
-			'structureId',
-			'title',
-			'trashed',
-			'type',
-			'unique',
-			'uri',
-			'with',
-			'withStructure'
-		];
-		$elementQuery = new LinkVaultDownloadQuery(LinkVaultDownload::class, []);
-		$criteriaAttributes = array_diff($elementQuery->criteriaAttributes(), $omittedCriteria);
-		$customFields = array_keys($this->plugin->customFields->fetchAllCustomFields('fieldName'));
-		$criteriaAttributes = array_merge($criteriaAttributes, $customFields);
-		sort($criteriaAttributes);
-		$options = [];
-		foreach($criteriaAttributes as $attr) {
-			$options[$attr] = Inflector::camel2words($attr);
+		$criteria = [];
+		foreach($input as &$filter) {
+			$fieldHandle = $filter['fieldHandle'] ?? null;
+			$filterType = $filter['filterType'] ?? null;
+			$value = $filter['value'] ?? null;
+			$newCriteria = $this->fieldCriteria($fieldHandle, $filterType, $value);
+			// Combine criteria for the same field into an array.
+			if ( isset($criteria[$fieldHandle]) ) {
+				// It might already be an "and where" array. Append this criteria to it.
+				if ( is_array($criteria[$fieldHandle]) ) {
+					$criteria[$fieldHandle][] = $newCriteria;
+				// Convert criteria into an array "and where" condition.
+				} else {
+					$originalCriteria = $criteria[$fieldHandle];
+					// We don't need the field key again. Just get the array value.
+					$criteriaOnly = end($newCriteria);
+					$criteria[$fieldHandle] = ['and', $originalCriteria, $criteriaOnly];
+				}
+			// No criteria was previously defined for this field so add it as usual.
+			} else {
+				$criteria = array_merge($criteria, $newCriteria);
+			}
 		}
+		return $criteria;
+	}
+
+	/**
+	 * This method returns an array of filter type options for a given field
+	 * based on its field handle. Once the field is fetched, it is actually the
+	 * "type" that determines what the filter options are.
+	 * @param string $handle
+	 * @param bool $asHtml
+	 * @param string $selectedValue
+	 * @return array|string
+	 */
+	public function getFilterOptionsByFieldHandle($handle, $asHtml=true, $selectedValue=null)
+	{
+		// Initialize the possible return value.
+		$options = [];
+		$optionsHtml = '';
+		$fields = [
+			'id' => ['is equal to', 'is greater than', 'is less than'],
+			'elementId' => ['is equal to', 'is greater than', 'is less than'],
+			'userId' => ['is equal to', 'is greater than', 'is less than'],
+			'assetId' => ['is equal to', 'is greater than', 'is less than'],
+			'type' => ['is equal to'],
+			's3Bucket' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'googleBucket' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'dirName' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'fileName' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'downloadAs' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'zipName' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'isUrl' => ['is equal to'],
+			'remoteIp' => ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'],
+			'dateCreated' => ['is greater than', 'is less than', 'is equal to', 'is empty', 'is not empty'],
+			'dateUpdated' => ['is greater than', 'is less than', 'is equal to', 'is empty', 'is not empty'],
+		];
+		$customFields = $this->plugin->customFields->fetchAllCustomFields('fieldName');
+		// Add the custom fields to the array with options appropriate for their type.
+		foreach($customFields as $customFieldHandle => &$customField) {
+			if ( in_array($customField->fieldType, ['varchar(250)', 'text']) ) {
+				$fields[$customFieldHandle] = ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'];
+			} else {
+				$fields[$customFieldHandle] = ['is equal to', 'is greater than', 'is less than'];
+			}
+		}
+		$options = $fields[$handle] ?? ['contains', 'starts with', 'ends with', 'is equal to', 'is empty', 'is not empty'];
+		if ( !empty($options) && $asHtml === true ) {
+			foreach($options as &$option) {
+				$selected = ($selectedValue && $selectedValue === $option) ? 'selected="selected"' : '';
+				$optionsHtml .= "<option value='{$option}' {$selected} >{$option}</option>";
+			}
+		}
+		$returnValue = ($asHtml === true) ? $optionsHtml : $options;
+		return $returnValue;
+	}
+
+	/**
+	 * This method attempts to fetch a filterable Link Vault Download record field's
+	 * available value options. There shouldn't be many fields that need this.
+	 * @param string $handle
+	 * @return array
+	 */
+	public function getFieldOptionsByFieldHandle($handle): array
+	{
+		$fieldsWithOptions = [
+			'isUrl' => [
+				1 => 'Yes',
+				0 => 'No'
+			],
+		];
+		$options = $fieldsWithOptions[$handle] ?? [];
 		return $options;
 	}