+++ src/variables/LinkVaultVariable.php
@@ -4,6 +4,7 @@ namespace Masuga\LinkVault\variables;
 
 use Craft;
 use Masuga\LinkVault\LinkVault;
+use Masuga\LinkVault\elements\db\LinkVaultCustomFieldQuery;
 use Masuga\LinkVault\elements\db\LinkVaultDownloadQuery;
 use Masuga\LinkVault\elements\db\LinkVaultReportQuery;
 
@@ -134,6 +135,17 @@ class LinkVaultVariable
 	}
 
 	/**
+	 * This template variable constructs and configures a LinkVaultCustomFieldQuery
+	 * then returns it.
+	 * @param array $criteria
+	 * @return LinkVaultCustomFieldQuery
+	 */
+	public function customFields($criteria=[]): LinkVaultCustomFieldQuery
+	{
+		return $this->plugin->customFields->customFieldsQuery($criteria);
+	}
+
+	/**
 	 * This template variable returns an array of leech attempt records based on
 	 * the given criteria.
 	 * @param array $criteria
@@ -163,6 +175,40 @@ class LinkVaultVariable
 	public function reports($criteria=[]): LinkVaultReportQuery
 	{
 		return $this->plugin->reports->reports($criteria);
+	}
+
+	/**
+	 * This method fetches the array of filter types based on a supplied field
+	 * handle. It also allows for an optional selected value parameter.
+	 * @param string $fieldHandle
+	 * @param string $selected
+	 * @return string
+	 */
+	public function fieldFilterOptions($fieldHandle, $selected=null): string
+	{
+		return $this->plugin->reports->getFilterOptionsByFieldHandle($fieldHandle, true, $selected);
+	}
+
+	/**
+	 * This method converts an array of filters form inputs into Craft element
+	 * query criteria.
+	 * @param array $criteria
+	 * @return array
+	 */
+	public function formatFilterCriteria($criteria): array
+	{
+		return $this->plugin->reports->formatCriteria($criteria);
+	}
+
+	/**
+	 * This template variable returns an associative array of field option values
+	 * and labels for a given field handle.
+	 * @param string $fieldHandle
+	 * @return array
+	 */
+	public function fieldOptions($fieldHandle): array
+	{
+		return $this->plugin->reports->getFieldOptionsByFieldHandle($fieldHandle);
 	}
 
 }